copilot/requirements.txt
pandas==2.2.3
numpy==1.26.4
PyYAML==6.0.2
streamlit==1.41.1
plotly==5.24.1
requests==2.32.3
pyarrow==18.1.0
